Rename encoding source files; Move SymbolTable into appropriate location
