+++ src/System.Text.Primitives/System/Text/Encoders/Ascii/Ascii_casing.cs
