+++ src/System.Text.Primitives/System/Text/Encoders/EncodingHelper.cs
@@ -3,7 +3,7 @@
 
 using System.Runtime.CompilerServices;
 
-namespace System.Text
+namespace System.Text.Encoders
 {
     internal static class EncodingHelper
     {
