+++ src/System.Text.Primitives/System/Text/Encoders/Utf16.cs
