+++ src/System.Text.Primitives/System/Text/Encoders/Utf32.cs
