+++ src/System.Text.Primitives/System/Text/ParsingTrie.cs
