+++ src/System.Text.Primitives/System/Text/SymbolTable.Symbol.cs
