+++ src/System.Text.Primitives/System/Text/SymbolTable.Utf16.cs
