+++ src/System.Text.Primitives/System/Text/SymbolTable.Utf8.cs
