+++ src/System.Text.Primitives/System/Text/SymbolTable.cs
